[ECO-5458] Moved LiveCounter and LiveMap to type-specific packages
- Moved LiveCounter.java to type/counter package for better organization
- Moved LiveMap.java to type/map package for better organization
- Updated import references in LiveObjects and ObjectsSubscription
