+++ lib/src/main/java/io/ably/lib/objects/ObjectsSubscription.java
@@ -11,12 +11,14 @@ package io.ably.lib.objects;
  * s.unsubscribe();
  * }
  * </pre>
+ * Spec: RTLO4b5
  */
 public interface ObjectsSubscription {
     /**
      * This method should be called when the subscription is no longer needed,
      * it will make sure no further events will be sent to the subscriber and
      * that references to the subscriber are cleaned up.
+     * Spec: RTLO4b5a
      */
     void unsubscribe();
 }
