+++ lib/src/main/java/io/ably/lib/objects/type/map/LiveMap.java
@@ -1,5 +1,6 @@
-package io.ably.lib.objects;
+package io.ably.lib.objects.type.map;
 
+import io.ably.lib.objects.ObjectsCallback;
 import org.jetbrains.annotations.Blocking;
 import org.jetbrains.annotations.NonBlocking;
 import org.jetbrains.annotations.Contract;
@@ -13,7 +14,7 @@ import java.util.Map;
  * The LiveMap interface provides methods to interact with a live, real-time map structure.
  * It supports both synchronous and asynchronous operations for managing key-value pairs.
  */
-public interface LiveMap {
+public interface LiveMap extends LiveMapChange {
 
     /**
      * Retrieves the value associated with the specified key.
